refactor(diff): deduplicate hunkToHeader and move runtime helpers to helpers.ts
Address review feedback on PR #269:
- Extract duplicated hunkToHeader from multibuffer.ts and controller.ts
  into a shared src/diff/helpers.ts module
- Use HunkHeader as the return type instead of an anonymous inline type
- Move formatHunkHeader from types.ts (type-only module) to helpers.ts
