+++ src/diff/types.ts
@@ -63,17 +63,3 @@ export interface HunkHeader {
   readonly context?: string;
 }
 
-/**
- * Format a HunkHeader into the standard unified diff header string.
- * Example: "@@ -10,5 +12,7 @@ function handleClick()"
- */
-export function formatHunkHeader(header: HunkHeader): string {
-  const oldPart = header.oldCount === 1
-    ? `${header.oldStart}`
-    : `${header.oldStart},${header.oldCount}`;
-  const newPart = header.newCount === 1
-    ? `${header.newStart}`
-    : `${header.newStart},${header.newCount}`;
-  const contextPart = header.context ? ` ${header.context}` : "";
-  return `@@ -${oldPart} +${newPart} @@${contextPart}`;
-}
